# Prototype for NonEmptyChunk