--- a/core-tests/shared/src/test/scala/zio/ChunkBufferSpec.scala
+++ b/core-tests/shared/src/test/scala/zio/ChunkBufferSpec.scala
@@ -396,20 +396,20 @@ object ChunkBufferSpec extends ZIOBaseSpec {
   )
 
   final def byteChunk(bytes: Byte*): Chunk[Byte] =
-    Chunk(bytes: _*)
+    Chunk.fromIterable(bytes)
 
   final def charChunk(chars: Char*): Chunk[Char] =
-    Chunk(chars: _*)
+    Chunk.fromIterable(chars)
 
   final def doubleChunk(doubles: Double*): Chunk[Double] =
-    Chunk(doubles: _*)
+    Chunk.fromIterable(doubles)
 
   final def floatChunk(floats: Float*): Chunk[Float] =
-    Chunk(floats: _*)
+    Chunk.fromIterable(floats)
 
   final def longChunk(longs: Long*): Chunk[Long] =
-    Chunk(longs: _*)
+    Chunk.fromIterable(longs)
 
   final def shortChunk(shorts: Short*): Chunk[Short] =
-    Chunk(shorts: _*)
+    Chunk.fromIterable(shorts)
 }