Following a discussion on Discord with @adamgfraser 
This PR is to add a `NonEmptyChunk[A]` to ZIO.

A `Chunk[A]` would be either an `Empty` or a `NonEmptyChunk[A]`.

Some methods are being overloaded in the `trait NonEmptyChunk[A]`, some are moving to an `Ops`.

Design decisions :
* When possible, return a `NonEmptyChunk[A]`
* All the subtype of `Chunk[A]`, except `Empty`, extend `NonEmptyChunk[A]`
* Minimize the change to existing code
* Complete compatibility with Scala 2.11 is optional (overloaded method are not working well in 2.11)

So far it seems possible! 😃