@ahoy-jon 2020-03-31
@adamgfraser if you can take a look to check it is going in the right direction!

On my list of things to add : 
- there are a couple of tests to add (reduce, flatMap, ... )
- performance tests