@jdegoes 2020-03-31
Organization-wise: think this should be `Chunk.NonEmpty` to mirror `Chunk.Empty`.