@ahoy-jon 2020-04-01
@jdegoes @adamgfraser Thanks you for the code review!

I have a lot to clean before the next review.
- 0 asInstanceOf
- reorder the methods in NonEmpty
- ...

As for the performance, it is similar to the previous version (thanks @Vilkina for quick help).

- Old Version
```scala
[info] Benchmark                     (size)  Mode  Cnt      Score      Error  Units
[info] MixedChunkBenchmarks.find       1000  avgt    5     64.854 ±    0.452  ns/op
[info] MixedChunkBenchmarks.flatMap    1000  avgt    5  80139.933 ± 3931.659  ns/op
[info] MixedChunkBenchmarks.fold       1000  avgt    5  22705.831 ± 1053.015  ns/op
[info] MixedChunkBenchmarks.foldM      1000  avgt    5  31908.680 ± 3762.721  ns/op
[info] MixedChunkBenchmarks.map        1000  avgt    5  26212.104 ± 2510.529  ns/op
[info] MixedChunkBenchmarks.mapM       1000  avgt    5  28448.603 ± 3103.181  ns/op
zio/master
```

- New Version
```scala
[info] Benchmark                     (size)  Mode  Cnt      Score      Error  Units
[info] MixedChunkBenchmarks.find       1000  avgt    5     61.773 ±    0.425  ns/op
[info] MixedChunkBenchmarks.flatMap    1000  avgt    5  80019.958 ±  201.439  ns/op
[info] MixedChunkBenchmarks.fold       1000  avgt    5  24802.410 ± 2749.465  ns/op
[info] MixedChunkBenchmarks.foldM      1000  avgt    5  29257.296 ± 1952.077  ns/op
[info] MixedChunkBenchmarks.map        1000  avgt    5  24998.218 ±  865.577  ns/op
[info] MixedChunkBenchmarks.mapM       1000  avgt    5  29055.468 ± 1118.978  ns/op
nonEmptyChunk
```

I don't find the difference conclusive, so more benchmarks!