@bernit77 2020-04-02
working with @ahoy-jon  , here are the performance test on my desktop 
with the command ```./sbt 'benchmarks/jmh:run  -i 15 -wi 15 -f1 -t2 ".*ChunkBenchmarks.*"'```
- nonEmptyChunk
```scala
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                     (size)  Mode  Cnt       Score      Error  Units
[info] MixedChunkBenchmarks.find       1000  avgt   15      81,038 ±    0,483  ns/op
[info] MixedChunkBenchmarks.flatMap    1000  avgt   15  110688,064 ± 1006,285  ns/op
[info] MixedChunkBenchmarks.fold       1000  avgt   15   36156,638 ±  178,179  ns/op
[info] MixedChunkBenchmarks.foldM      1000  avgt   15   38390,781 ±  242,665  ns/op
[info] MixedChunkBenchmarks.map        1000  avgt   15   36968,630 ±  247,083  ns/op
[info] MixedChunkBenchmarks.mapM       1000  avgt   15   38547,848 ±  195,323  ns/op
[success] Total time: 1885 s (31:25), completed 2 avr. 2020 11:09:29
ahoy-jon/master
```
- master
```scala
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                     (size)  Mode  Cnt       Score      Error  Units
[info] MixedChunkBenchmarks.find       1000  avgt   15      81,806 ±    1,835  ns/op
[info] MixedChunkBenchmarks.flatMap    1000  avgt   15  114643,764 ± 1404,866  ns/op
[info] MixedChunkBenchmarks.fold       1000  avgt   15   27936,776 ±  318,577  ns/op
[info] MixedChunkBenchmarks.foldM      1000  avgt   15   34992,621 ±  274,165  ns/op
[info] MixedChunkBenchmarks.map        1000  avgt   15   34528,973 ±  173,703  ns/op
[info] MixedChunkBenchmarks.mapM       1000  avgt   15   37150,904 ±  246,418  ns/op
[success] Total time: 1865 s (31:05), completed 2 avr. 2020 11:45:50
zio/ef148c12
```