@ahoy-jon 2020-04-02
Thanks!

I will take a look at `fold`.